10.18 몬스터 애니메이션 적용

+++ TeamByte/Assets/_Script/Monster/MonsterData.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Monster Data", menuName = "ScriptableObject/Monster Data Asset", order = 2)]
 public class MonsterData : ScriptableObject
 {
-    public float m_iAttackDamage;
-    public float m_iHP;
+    public int m_iAttackDamage;
+    public int m_iHP;
     public float m_iSpeed;
     public float m_iAttackRange;
     public MonsterData()
